+++ src/query/service/src/pipelines/processors/transforms/aggregator/transform_group_by_partial.rs
@@ -217,8 +217,8 @@ impl<Method: HashMethodBounds> AccumulatingTransform for TransformPartialGroupBy
             HashTable::PartitionedHashTable(v) => {
                 info!(
                     "Processed {} rows, allocated {} memory while in group by.",
-                    convert_number_size(v.len()),
-                    convert_byte_size(v.allocated_bytes())
+                    convert_number_size(v.len() as f64),
+                    convert_byte_size(v.allocated_bytes() as f64)
                 );
                 let _ = v.hashtable.unsize_key_size();
                 let cells = PartitionedHashTableDropper::split_cell(v);
